tests on delete events

+++ src/handlers/delete-events.js
@@ -1,4 +1,4 @@
-import { deleteEvent } from "../data-layer/events.data-mapper.js";
+import { deleteEvent, getEvent } from "../data-layer/events.data-mapper.js";
 import { wrapHandler } from "../utilities/handlerWrapper.js";
 import { getUserId } from "../utilities/request.js";
 import { assembleHandleResponse } from "../utilities/response.js";
@@ -19,15 +19,19 @@ const parseParams = (event) => {
   validate(params, paramSchema, "param");
 
   return {
-    StartDateTime: new Date(params.StartDateTime)
-  }
-}
+    StartDateTime: new Date(params.StartDateTime),
+  };
+};
 
 export const handler = wrapHandler(async (event) => {
   const params = parseParams(event);
   const userId = getUserId();
 
-  const r = await deleteEvent(userId, params.StartDateTime)
-  
-  return assembleHandleResponse(200, {});
+  if (!(await getEvent(userId, params.StartDateTime))) {
+    return assembleHandleResponse(404, { message: "Event did not found!" });
+  }
+
+  await deleteEvent(userId, params.StartDateTime);
+
+  return assembleHandleResponse(200, { message: "Item deleted successfully!" });
 });
